update BeforeTargets name for new Xamarin.VS projects
Hey guys, I'm not an msbuild expert, but I found that this awesome XCC preprocessor had stopped working when I included it in a brand new Xamarin.Forms project I created with Visual Studio 2015 (Update 2). I was actually heartbroken, because it is one of THE most useful utilities I've ever come across for the ignorable xaml intellisense attributes.

I did a little digging, and it looks like sometime in the last 9 months or so, they've changed the name of the target task from `MarkupCompilePass1` to `XamlMarkupCompilePass1`.

I tweaked the `xcc.targets` file in my nuget packages folder to use this new name (as per this PR) and it worked!

Sadly I am not an msbuild expert, so I do not know if this will break previous versions, and I do not know how to make it do a check for multiple targets (e.g. something like `BeforeTargets="MarkupCompilePass1,XamlMarkupCompilePass1"` or what order they should be in, or even if it can be be done this way). I hope you can help out with this, because there's a few sad xamarin forms devs out there right now.

Thanks for building this thing! It is SUCH a time saver! :thumbsup: